+++ src/ui/views/standard.jsx
@@ -44,6 +44,11 @@ export function standard(dataset){
         cssClasses.push('enlighter-overflow-scroll');
     }
 
+    // limit vertical height ?
+    if (options.collapse === true){
+        cssClasses.push('enlighter-collapse');
+    }
+
     // additional css classes set ?
     if (options.cssClasses.length > 0){
         cssClasses.push(...options.cssClasses);